chore: move snippet processing code to separate module
